chore(tests): create test_config_entry.py
